All supported languages are fully implemented, tests are in progress

+++ Java/src/org/hpms/mw/distcrud/Network.java
@@ -6,7 +6,7 @@ import java.net.NetworkInterface;
 
 public class Network {
 
-   public static IParticipant join( byte id, InetSocketAddress addr, NetworkInterface intrfc ) throws IOException {
+   public static IParticipant join( int id, InetSocketAddress addr, NetworkInterface intrfc ) throws IOException {
       return new ParticipantImpl( id, addr, intrfc );
    }
 }
