+++ Java/test/tests/person/Test_001.java
@@ -14,8 +14,8 @@ public class Test_001 implements Settings {
       if( Tests.checkNetworkInterface( via )) {
          final InetSocketAddress p1 = new InetSocketAddress( MCAST_ADDR, PORT   );
          final InetSocketAddress p2 = new InetSocketAddress( MCAST_ADDR, PORT+1 );
-         new Publisher ((byte)1, p1, via, p2 );
-         new Subscriber((byte)2, p2, via, p1 );
+         new Publisher ( 1, p1, via, p2 );
+         new Subscriber( 2, p2, via, p1 );
       }
    }
 }
