+++ Java/test/tests/person/Test_006.java
@@ -14,7 +14,7 @@ public class Test_006 implements Settings {
       if( Tests.checkNetworkInterface( via )) {
          final InetSocketAddress p1 = new InetSocketAddress( MCAST_ADDR, PORT   );
          final InetSocketAddress p2 = new InetSocketAddress( MCAST_ADDR, PORT+1 );
-         new Publisher((byte)1, p1, via, p2 );
+         new Publisher( 1, p1, via, p2 );
       }
    }
 }
